Only send IV_NCP=2 when supporting AES GCM ciphers
A quirk from OpenVPN 2.4 is that the peer-info value IV_NCP=2 signals
the client supports (and wants to use) AES-128-GCM and AES-256-GCM.
Newer openvpn servers (>= 2.5) will look after IV_NCP=2 if IV_CIPHERS is not
present.

The IV_NCP=2 value could be removed if we do not desire to support
OpenVPN 2.4 (and other implementations with the same quirk).