Improve --topology implementation
In the server we only support --topology subnet. Instead of hardcoding
"topology subnet" in the push reply we get the value from our config. We
also assume --topology subnet if --topology is not present. This breaks
from OpenVPN<2.7.

Finally, document that sending --cipher in push reply is how cipher
negotiation is concluded.