Benchmark (#230)
The benchmark sets up a client and server internally and does the handshake until established (by way of hard coded parsing messages emitted from each peer). The benchmark tests the data encoding ("upload") and data decoding ("download") and does not exercise the control channel as part of the benchmark. The benchmark is further parameterized on the (data channel) cipher.