Control channel explicit exit notifications (#247)
Parse better control channel messages

Messages EXIT, RESTART and HALT are now better parsed. Messages RESTART
and HALT can carry an optional message separated by a comma, e.g.

    HALT,shutting down

The OpenVPN parser is very sloppy and only checks the prefixes "EXIT",
"HALT", and "RESTART". Thus "HALTTTTT" is interpreted by OpenVPN as a
HALT message. We choose not to be this sloppy with the parsing.

Co-authored-by: Hannes Mehnert <hannes@mehnert.org>